+++ src/main/java/ca/corefacility/bioinformatics/irida/ria/web/samples/SamplesAjaxController.java
@@ -8,11 +8,9 @@ import java.util.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.MessageSource;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.multipart.MultipartHttpServletRequest;
 
@@ -22,6 +20,7 @@ import ca.corefacility.bioinformatics.irida.model.sequenceFile.Fast5Object;
 import ca.corefacility.bioinformatics.irida.model.sequenceFile.SequenceFile;
 import ca.corefacility.bioinformatics.irida.model.sequenceFile.SequenceFilePair;
 import ca.corefacility.bioinformatics.irida.model.sequenceFile.SingleEndSequenceFile;
+import ca.corefacility.bioinformatics.irida.ria.web.samples.dto.SampleDetails;
 import ca.corefacility.bioinformatics.irida.service.GenomeAssemblyService;
 import ca.corefacility.bioinformatics.irida.service.SequencingObjectService;
 import ca.corefacility.bioinformatics.irida.service.sample.SampleService;
@@ -152,6 +151,20 @@ public class SamplesAjaxController {
 					.body(messageSource.getMessage("server.SampleFileUploader.error",
 							new Object[] { sample.getSampleName() }, locale));
 		}
+	}
+
+	/**
+	 * Get {@link Sample} details for a specific sample.
+	 *
+	 * @param id {@link Long} identifier for a sample.
+	 * @return {@link SampleDetails} for the {@link Sample}
+	 */
+	@RequestMapping(value = "/{id}/details", produces = MediaType.APPLICATION_JSON_VALUE)
+	@ResponseBody
+	public SampleDetails getSampleDetails(@RequestParam Long id) {
+		Sample sample = sampleService.read(id);
+		boolean modifiable = this.isSampleModifiable(sample);
+		return new SampleDetails(sample, modifiable);
 	}
 
 	/**
