+++ src/main/java/ca/corefacility/bioinformatics/irida/ria/web/samples/SamplesController.java
@@ -15,7 +15,6 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.MessageSource;
 import org.springframework.format.annotation.DateTimeFormat;
-import org.springframework.http.MediaType;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
@@ -41,7 +40,6 @@ import ca.corefacility.bioinformatics.irida.model.sequenceFile.SequencingObject;
 import ca.corefacility.bioinformatics.irida.model.sequenceFile.SingleEndSequenceFile;
 import ca.corefacility.bioinformatics.irida.model.user.User;
 import ca.corefacility.bioinformatics.irida.ria.web.BaseController;
-import ca.corefacility.bioinformatics.irida.ria.web.samples.dto.SampleDetails;
 import ca.corefacility.bioinformatics.irida.security.permissions.sample.UpdateSamplePermission;
 import ca.corefacility.bioinformatics.irida.service.GenomeAssemblyService;
 import ca.corefacility.bioinformatics.irida.service.ProjectService;
@@ -147,20 +145,6 @@ public class SamplesController extends BaseController {
 		model.addAttribute(MODEL_ATTR_ACTIVE_NAV, ACTIVE_NAV_DETAILS);
 		model.addAttribute(MODEL_ATTR_CAN_MANAGE_SAMPLE, isSampleModifiable(sample));
 		return SAMPLE_PAGE;
-	}
-
-	/**
-	 * Get {@link Sample} details for a specific sample.
-	 *
-	 * @param id {@link Long} identifier for a sample.
-	 * @return {@link SampleDetails} for the {@link Sample}
-	 */
-	@RequestMapping(value = "/samples", produces = MediaType.APPLICATION_JSON_VALUE)
-	@ResponseBody
-	public SampleDetails getSampleDetails(@RequestParam Long id) {
-		Sample sample = sampleService.read(id);
-		boolean modifiable = this.isSampleModifiable(sample);
-		return new SampleDetails(sample, modifiable);
 	}
 
 	/**
