+++ src/main/webapp/resources/js/apis/samples/samples.js
@@ -1,11 +1,11 @@
 import axios from "axios";
 import { setBaseUrl } from "../../utilities/url-utilities";
 
-const URL = setBaseUrl(`samples`);
+const URL = setBaseUrl(`ajax/samples`);
 
 export const getSampleDetails = async (id) => {
   try {
-    const { data } = await axios.get(`${URL}?id=${id}`);
+    const { data } = await axios.get(`${URL}/${id}/details`);
     return data;
   } catch (e) {
     return Promise.reject(e.response.data.error);
